Se añadió el controller Categoria.

+++ src/main/java/com/librofavoritobackend/model/Libro.java
@@ -17,7 +17,7 @@ public class Libro {
     private Long id;
     
     @Column(unique = true)
-    private int isbn;
+    private Long isbn;
     
     @Column(nullable = false)
     private String titulo;
